Makes code compatible with Scala 2.13
