+++ src/test/scala/higherkindness/skeuomorph/avro/AvroSpec.scala
@@ -61,8 +61,8 @@ class AvroSpec extends Specification with ScalaCheck {
         .and(sch.getNamespace should_== namespace.getOrElse(""))
         .and(sch.getDoc should_== doc.getOrElse(""))
         .and(
-          sch.getFields.asScala.toList.map(f => (f.name, f.doc)) should_== fields.map(f => (f.name, f.doc.getOrElse(""))
-          )
+          sch.getFields.asScala.toList.map(f => (f.name, f.doc)) should_== fields
+            .map(f => (f.name, f.doc.getOrElse("")))
         )
 
     case AvroF.TEnum(_, _, _, _, _) => true
